refactor: LineService는Repository를 통해 도메인만 다루도록 수정
