+++ src/main/java/wooteco/subway/domain/line/LineRepository.java
@@ -3,9 +3,9 @@ package wooteco.subway.domain.line;
 import java.util.List;
 
 public interface LineRepository {
-    List<Line> findAll();
-
     Line save(Line line, long upStationId, long downStationId, int distance);
+
+    List<Line> findAll();
 
     Line find(long id);
